A shopping search engine
This was built to truly combine textual features from web pages, and
visual features from image content for better image search in a vertical
search engine. Testing was done based on data which was collected from
Macy's website. I built a crawler for crawling data from Macy's website.
Java and Selenium was used for this purpose